## 권순우 ## <MainScene 및 GameObject 클래스 기능 추가> - 각자 폴더에 MainScene에서 Init할 함수를 추가해두었음 ㄴ 자신이 만든 클래스 테스트 하고 싶다하면 그곳에 new 할당 및 ObjectManager에 AddObject하면됌
- GameObject 클래스 간단한 수정 및 추가 되었음. 추가된 기능들의 사용법들은 주말에 수업할 예정

+++ TeamProject/TeamProject/TextFile.cpp
@@ -153,7 +153,7 @@ int TextReader::Int()
 float TextReader::Float()
 {
 	endIndex = tempStr.find(TextReader::spaceToken);
-	float returnValue = atof(tempStr.substr(0, endIndex).c_str());
+	float returnValue = (float)atof(tempStr.substr(0, endIndex).c_str());
 	tempStr.erase(0, endIndex + 1);
 
 	return returnValue;
